Add minimalistic symbolic framework
cc: @adamc @abergal @andres-erbsen

src/Minimalistic.v now contains a very minimal framework for reasoning
about computation "with high probability" in presence of limited
adversarial interaction. The approach is vaguely inspired by
[1](https://eprint.iacr.org/2016/069.pdf), but the differences are
perhaps more significant than the similarities.

1. As in [1], all randomess is pre-generated at once in the beginning of
the protocol and all interesting computation is pure.
2. There is no formalization of first order logic, and there is no
codification of models. Instead, Coq quantifiers are used to hide
information from the "symbolic" side. In particular, there no longer are
two models (symbolic and computational), there is now just the
computational model, but most proofs hide the majority of it by
quantifying over a large swath of the semantics.
3. There is no trace of the "execution" transition system framework from
previous computationally complete symbolic attacker works that also
appears in [1]. I am guessing that we are better off building our own
instead of replicating that. For example, we probably want to never
formalize the non-folded protocol representation and instead make
notation for writing the "folded" protocols in a human-readable way.
4. There are no length-indexed "sequences", saving us from dependent
type hell of the previous formalizations. Instead, the messages revealed
to the adversary.

In the bottom of the file, there is one "symbolic world" proof that
composes two axioms to achieve a result that needs both of them. And
there also is a not-yet-proven lemma for the soundness of one of those
axioms, and it looks reasonable.

I have not yet worked with axioms that require limiting the
computational power of the adversary, but I don't expect any additional
complications.

More concerningly, I have not checked whether inputs from the adversary
behave as we wish they did. Ideally it should just work out, I think I
put the right things in right places. But I am done working for today,
so that will be some other time.